refactor : 프로젝트 구조와 함수 분리
